% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIENtableCleanup.R
\name{BIENtableCleanup}
\alias{BIENtableCleanup}
\title{BIEN Table Cleanup}
\usage{
BIENtableCleanup(BIENtable)
}
\arguments{
\item{BIENtable}{A data.frame resulting from \code{\link{getBIENpoints}}.}
}
\value{
A data.frame tidied for compatibility with
\code{\link{getGBIFpoints}} results.
}
\description{
Forces occurrence table columns to a uniform
format. This is an internal function for use by \code{\link{occQuery}}.
}
\examples{
\dontrun{
temp <- getBIENpoints(taxon="Protea cynaroides", limit = NULL)
fitForUseTable <- BIENtableCleanup(temp$OccurrenceTable)
}

}
