% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFLogin.R
\docType{class}
\name{GBIFLogin-class}
\alias{GBIFLogin-class}
\alias{GBIFLogin}
\title{GBIFLogin Data Class}
\description{
A class for managing GBIF login data.
}
\section{Slots}{

\describe{
\item{\code{username}}{A vector of type character specifying a GBIF username.}

\item{\code{email}}{A vector of type character specifying the email associated with
a GBIF username.}

\item{\code{pwd}}{A vector of type character containing the user's password for
logging in to GBIF.}
}}

\examples{
\donttest{
GBIFLogin <- GBIFLoginManager(user = "occCiteTester",
     email = "****@yahoo.com",
     pwd = "12345")
}

}
