% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFLoginManager.R
\name{GBIFLoginManager}
\alias{GBIFLoginManager}
\title{GBIF Login Manager}
\usage{
GBIFLoginManager(user = NULL, email = NULL, pwd = NULL)
}
\arguments{
\item{email}{A vector of type character specifying the email associated with a GBIF username.}

\item{pwd}{A vector of type character containing the user's password for logging in to GBIF.}

\item{username}{A vector of type character specifying a GBIF username.}
}
\value{
An object of class \code{\link{GBIF}} containing the user's GBIF login data.
}
\description{
GBIF Login Manager
}
\examples{
## Inputting user particulars
myLogin <- GBIFLoginMaster(user = "theWoman", email = "ireneAdler@laScala.org", pwd = "sh3r");

## Can also be mined from your system environment
myLogin <- GBIFLoginMaster(user = NULL, email = NULL, pwd = NULL);

}
