% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifRetriever.R
\name{gbifRetriever}
\alias{gbifRetriever}
\title{Retrieve downloaded GBIF data sets}
\usage{
gbifRetriever(taxon = NULL)
}
\arguments{
\item{taxon}{A single species}
}
\value{
A list of lists containing (1) a data frame of occurrence
data; (2) GBIF search metadata for every GBIF download in the
specified directory. (3) the raw occurrence data from GBIF.
}
\description{
Searches for the most recent instance of a data set
 that has been previously downloaded from GBIF to a local machine
 and imports it into `occCite`. This is designed to be an internal
 function. It is usable on its own, but you must have already
 navigated to the folder that contains the download zip files
 from GBIF.
}
\keyword{internal}
