% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIENpoints.R
\name{getBIENpoints}
\alias{getBIENpoints}
\title{Download occurrence points from BIEN}
\usage{
getBIENpoints(taxon)
}
\arguments{
\item{taxon}{A single plant species or vector of plant species}
}
\value{
A list containing (1) a dataframe of occurrence data; (2) a list containing: i notes on usage, ii bibtex citations, and iii acknowledgement information; (3) a dataframe containing the raw results of a query to `BIEN::BIEN_occurrence_species()`.
}
\description{
Downloads occurrence points and useful related information for processing within other occCite functions
}
\examples{
\dontrun{
getBIENpoints(taxon="Acer rubrum", limit = NULL);
}
}
