% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occCite_plotting.R
\name{map.occCite}
\alias{map.occCite}
\title{Generating a map of downloaded points}
\usage{
map.occCite(occCiteData, cluster = FALSE)
}
\arguments{
\item{occCiteData}{An object of class \code{\link{occCiteData}} to
map.}

\item{cluster}{Logical; setting to `TRUE` turns on marker clustering.}
}
\value{
A GBIF download key, if one is available
}
\description{
Makes maps for each individual species in an occCite
data object.
}
\examples{
\donttest{
map.occCite(occCiteData, cluster = FALSE)
}

}
