% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occCitation.R
\name{occCitation}
\alias{occCitation}
\title{Occurrence Citations}
\usage{
occCitation(x = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{occCiteData}}}
}
\value{
A data frame with citations information for occurrences
}
\description{
Harvests citations for occurrence data
}
\examples{
\dontrun{
data(myOccCiteObject)
myCitations <- occCitation(x = myOccCiteObject)
cat(paste(myCitations$Citation, " Accessed via ",
     myCitations$occSearch, " on ", myCitations$`Accession Date`, "."),
     sep = "\n")
}
}
