% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occCiteCitation.R
\docType{class}
\name{occCiteCitation-class}
\alias{occCiteCitation-class}
\alias{occCiteCitation}
\title{occCite Citation Class}
\description{
A class for managing citations generated from occCite
queries.
}
\section{Fields}{

\describe{
\item{\code{occCitationResults}}{The results of performing
\code{\link{occCitation}} on a \code{\link{occCiteData}} object,
stored as a named list, each of the items named after a searched
taxon and containing a data frame with occurrence information.}
}}


