% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occCitePrint.R
\name{print.occCiteCitation}
\alias{print.occCiteCitation}
\title{Print occCite citation object}
\usage{
\method{print}{occCiteCitation}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{occCiteCitation}}}

\item{...}{Additional arguments affecting how the formatted
citation document is produced}

\item{bySpecies}{Logical; If `TRUE`, formatted citations will
be returned in chunks headed by the species name. If `FALSE`,
a single citation string for all species is returned.}
}
\value{
A text string with formatted citatations
}
\description{
Prints formatted citations for occurrences
}
\examples{

# Print citations for all species together
data(myOccCiteObject)
print(myOccCiteObject, bySpecies = F)

# Print citations for each species individually
data(myOccCiteObject)
print(myOccCiteObject, bySpecies = T)

}
