% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occSum.R
\name{summary.occCiteData}
\alias{summary.occCiteData}
\title{Summary for occCite data objects}
\usage{
\method{summary}{occCiteData}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{occCiteData}}}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
A dataframe with citations information for occurrences
}
\description{
Displays a summary of relevant stats about a query
}
\examples{

\dontrun{
myOcciteObject <- occQuery(x = myOccCiteObject,
         datasources = c("gbif", "bien"),
         GBIFLogin = myLogin,
         limit = NULL,
         GBIFDownloadDirectory = "./Desktop"
         loadLocalGBIFDownload = F);
summary(myOcciteObject);
}

}
