% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifRetriever.R
\name{gbifRetriever}
\alias{gbifRetriever}
\title{Retrieve downloaded GBIF datasets}
\usage{
gbifRetriever(taxon = NULL)
}
\arguments{
\item{taxon}{A single species}
}
\value{
A list of lists containing (1) a dataframe of occurrence
data; (2) GBIF search metadata for every GBIF download in the
specified directory.
}
\description{
Searches for the most recent instance of a dataset
 that has been previously downloaded from GBIF to a local machine
  and imports it into occCite
}
\examples{
\dontrun{
gbifRetriever(GBIFDownloadDirectory = NULL);
}

}
