% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGBIFpoints.R
\name{getGBIFpoints}
\alias{getGBIFpoints}
\title{Download occurrence points from GBIF}
\usage{
getGBIFpoints(taxon, GBIFLogin = GBIFLogin,
  GBIFDownloadDirectory = GBIFDownloadDirectory, limit = NULL)
}
\arguments{
\item{taxon}{A single species}

\item{GBIFLogin}{An object of class \code{\link{GBIFLogin}} to log in to GBIF to begin the download.}

\item{GBIFDownloadDirectory}{An optional argument that specifies the local directory where GBIF downloads will be saved. If this is not specified, the downloads will be saved to your current working directory.}
}
\value{
A list containing (1) a dataframe of occurrence data; (2) GBIF search metadata
}
\description{
Downloads occurrence points and useful related information for processing within other occCite functions
}
\examples{
getGBIFpoints(taxon="Gadus morhua", GBIFLogin = myGBIFLogin, GBIFDownloadDirectory = NULL);

}
