% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occCitation.R
\name{occCitation}
\alias{occCitation}
\title{Occurrence Citations}
\usage{
occCitation(x = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{occCiteData}}}
}
\value{
An object of class \code{\link{occCiteCitation}}. It is
a named list of the same length as the number of species
included in your \code{\link{occCiteData}} object. Each item
in the list has citation information for occurrences.
}
\description{
Harvests citations for occurrence data
}
\examples{
\dontrun{
data(myOccCiteObject)
myCitations <- occCitation(x = myOccCiteObject)
}
}
