% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occCitation.R
\name{occCitation}
\alias{occCitation}
\title{Occurrence Citations}
\usage{
occCitation(x = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{occCiteData}}}
}
\value{
A dataframe with citations information for occurrences
}
\description{
Harvests citations for occurrence data
}
\examples{

\dontrun{
myCitations <- occCitation(x = myoccCiteObject);
}

}
