% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occCitation.R
\name{occCitation}
\alias{occCitation}
\title{Occurrence Citations}
\usage{
occCitation(x = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{occCiteData}}}
}
\value{
A dataframe with citations information for occurrences
}
\description{
Harvests citations from GBIF for occurrence data
}
\examples{
myCitations <- occCitation(x = myoccCiteObject);

}
