% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occCitePrint.R
\name{print.occCiteCitation}
\alias{print.occCiteCitation}
\title{Print occCite citation object}
\usage{
\method{print}{occCiteCitation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{occCiteCitation}}}

\item{...}{Additional arguments affecting how the formatted
citation document is produced}
}
\value{
A text string with formatted citations
}
\description{
Prints formatted citations for occurrences and main
packages used (i.e. base, occCite, rgbif, and/or BIEN).
}
\examples{

# Print citations for all species together
data(myOccCiteObject)
print(myOccCiteObject)

# Print citations for each species individually
data(myOccCiteObject)
print(myOccCiteObject, bySpecies = TRUE)
}
