% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occSum.R
\name{summary.occCiteData}
\alias{summary.occCiteData}
\title{Summary for occCite data objects}
\usage{
\method{summary}{occCiteData}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{occCiteData}}}

\item{...}{Additional arguments affecting the summary produced}
}
\description{
Displays a summary of relevant stats about a query
}
\examples{

data(myOccCiteObject)
summary(myOccCiteObject)

}
