% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabGBIF.R
\name{tabGBIF}
\alias{tabGBIF}
\title{GBIF Table}
\usage{
tabGBIF(GBIFresults, taxon)
}
\arguments{
\item{GBIFresults}{The results of a GBIF search that will be tabulated into a common format for an occCite object.}

\item{taxon}{A single species name, for tracing/error checking purposes only.}
}
\value{
A list of lists containing (1) a dataframe of occurrence data; (2) GBIF search metadata for every GBIF download in the specified directory.
}
\description{
Internal function--imports results from `occ_download_get()` and processes them into a table for an `occCiteData` object.
}
\examples{
\dontrun{
res <- rgbif::occ_download_get(key=downloadKey, overwrite=TRUE, file.path(getwd(), taxon));
tG <- rgbif::occ_download_import(res);
tabGBIF(GBIFresults = tG);
}

}
