% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonRectification.R
\name{taxonRectification}
\alias{taxonRectification}
\title{Taxon Rectification}
\usage{
taxonRectification(taxName = NULL, datasources = NULL)
}
\arguments{
\item{taxName}{A string that, ideally, is a taxonomic name}

\item{datasources}{A vector of taxonomic datasources implemented in \code{\link{gnr_resolve}}. See the \href{http://gni.globalnames.org/}{Global Names List} for more information.}
}
\value{
A string with the closeset match according to \code{\link{gnr_resolve}}, and a list of taxonomic datasources that contain the matching name.
}
\description{
An internal helper function that takes an input taxonomic name, checks against taxonomic database, returns vector for use in database queries, as well as warnings if the name is invalid.
}
\examples{
#Inputting a taxonomic name and specifying what taxonomic sources you want to search
studyTaxonList(x = "Buteo buteo hartedi", datasources = c('NCBI', 'EOL'));

}
